requirements.txt
camelot-py==1.0.0
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
Deprecated==1.2.18
et_xmlfile==2.0.0
lxml==5.3.2
numpy==2.2.4
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pdfminer.six==20250327
pillow==11.1.0
pycparser==2.22
pypdf==5.4.0
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tabulate==0.9.0
tzdata==2025.2
wrapt==1.17.2
